+++ src/test/scala/com/github/vitalsoftware/scalaredox/RedoxTest.scala
@@ -2,10 +2,11 @@ package com.github.vitalsoftware.scalaredox
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
-import com.github.vitalsoftware.scalaredox.client.{ RedoxClient, RedoxResponse }
+import com.github.vitalsoftware.scalaredox.client.{ ClientConfig, RedoxClient, RedoxResponse, RedoxTokenManager }
 import com.github.vitalsoftware.util.JsonImplicits.JsValueExtensions
 import com.typesafe.config.ConfigFactory
 import play.api.libs.json.{ JsError, Json, Reads }
+import play.api.libs.ws.ahc.StandaloneAhcWSClient
 
 import scala.concurrent.{ Await, Future }
 import scala.concurrent.duration._
@@ -15,10 +16,13 @@ import scala.concurrent.duration._
  */
 trait RedoxTest {
 
-  val conf = ConfigFactory.load("resources/reference.conf")
-  val system = ActorSystem("redox-test")
-  val materializer = ActorMaterializer()(system)
-  val client = new RedoxClient(conf, system, materializer)
+  val conf = ClientConfig(ConfigFactory.load("resources/reference.conf"))
+  val httpClient = StandaloneAhcWSClient()
+  implicit val system = ActorSystem("redox-test")
+  implicit val materializer = ActorMaterializer()(system)
+
+  val tokenManager = new RedoxTokenManager(httpClient, conf.baseRestUri)
+  val client = new RedoxClient(conf, httpClient, tokenManager)
   val timeout: FiniteDuration = 20.seconds
 
   // Validate raw a raw JSON string, throwing a RuntimeException which will output detail to Specs2/Test console